requirements/conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
appdirs=1.4.4=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
beautifulsoup4=4.12.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cbor=1.0.0=pypi_0
cbor2=5.4.6=pypi_0
certifi=2022.12.7=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
cryptography=39.0.1=py39h9ce1e76_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
datasets=2.13.1=pypi_0
dill=0.3.6=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
faiss-cpu=1.7.3=pypi_0
faiss-gpu=1.7.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py39h06a4308_0
flatten-json=0.1.13=pypi_0
flit-core=3.8.0=py39h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=pypi_0
fsspec=2023.6.0=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
huggingface-hub=0.13.3=pypi_0
idna=3.4=py39h06a4308_0
ijson=3.2.0.post0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ir-datasets=0.5.4=pypi_0
jinja2=3.1.2=py39h06a4308_0
joblib=1.2.0=pypi_0
jpeg=9e=h5eee18b_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.6.0.25=0
libcurand=10.3.2.56=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
llvmlite=0.39.1=pypi_0
lxml=4.9.2=pypi_0
lz4=4.3.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.1=py39h7f8727e_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py39h06a4308_0
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py39h06a4308_1
nltk=3.8.1=pypi_0
numba=0.56.4=pypi_0
numpy=1.23.5=py39h14f4228_0
numpy-base=1.23.5=py39h31eccc5_0
onnx=1.14.0=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
orjson=3.8.9=pypi_0
packaging=23.0=pypi_0
pandas=2.0.0=pypi_0
pathtools=0.1.2=pypi_0
pillow=9.4.0=py39h6a678d5_0
pip=23.0.1=py39h06a4308_0
protobuf=4.23.2=pypi_0
psutil=5.9.5=pypi_0
pyarrow=12.0.1=pypi_0
pyautocorpus=0.1.9=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.14.0=pypi_0
pyopenssl=23.0.0=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.16=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pypi_0
pytorch=2.0.0=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_3
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
rank-bm25=0.2.2=pypi_0
ranx=0.3.6=pypi_0
readline=8.2=h5eee18b_0
regex=2023.3.23=pypi_0
requests=2.28.1=py39h06a4308_1
rich=13.3.3=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.97=pypi_0
sentry-sdk=1.25.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.6.3=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
soupsieve=2.4=pypi_0
sqlite=3.41.1=h5eee18b_0
sympy=1.11.1=py39h06a4308_0
tabulate=0.9.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.2=pypi_0
torchaudio=2.0.0=py39_cu117
torchtriton=2.0.0=py39
torchvision=0.15.0=py39_cu117
tqdm=4.65.0=py39hb070fc8_0
transformers=4.27.4=pypi_0
trec-car-tools=2.6=pypi_0
typing_extensions=4.4.0=py39h06a4308_0
tzdata=2023.3=pypi_0
unlzw3=0.2.2=pypi_0
urllib3=1.26.15=py39h06a4308_0
wandb=0.15.3=pypi_0
warc3-wet=0.2.3=pypi_0
warc3-wet-clueweb09=0.2.5=pypi_0
wheel=0.38.4=py39h06a4308_0
xxhash=3.2.0=pypi_0
xz=5.2.10=h5eee18b_1
yacs=0.1.8=pypi_0
yarl=1.9.2=pypi_0
zlib=1.2.13=h5eee18b_0
zlib-state=0.1.5=pypi_0
zstd=1.5.4=hc292b87_0
